blog/api/serializers.py
from django.contrib.auth.models import User, Group
from rest_framework import serializers
from blog.models import Comment

class CommentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Comment
        fields = ['id','post','author','text']
        

