requirements.txt
appdirs==1.4.4
certifi==2020.12.5
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.7
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
pipenv==2020.11.15
psycopg2-binary==2.8.6
python-memcached==1.59
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.6
virtualenv-clone==0.5.4
whitenoise==5.2.0
